Intento de arreglar bug de documentacion en conf.py

+++ docs/conf.py
@@ -38,8 +38,10 @@ if not on_rtd:  # only import and set the theme if we're building docs locally
     html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
 else:
-    os.rename(os.path.join(project_root, 'detector_aedes', 'config_sample.yml'),
-              os.path.join(project_root, 'detector_aedes', 'config.yml'))
+    config_file_path = os.path.join(project_root, 'detector_aedes', 'config.yml')
+    if not os.path.isfile(config_file_path):
+        os.rename(os.path.join(project_root, 'detector_aedes', 'config_sample.yml'),
+                  config_file_path)
     html_theme = 'default'
 
 import detector_aedes
